+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## 1.0.0 (unreleased)
 - Première version "stable".
 - Met à jour les dépendances de développement.
+- Ajoute des fonctions du core de SASS dans les exceptions de la règles 
+  `scss/at-function-named-arguments` vu que les arguments nommés dans ce
+  cas servent d'indicateurs.
 
 ## 0.2.0 (2020-07-17)
 - Autorise les préfixes `_` dans les variables.